requirements.txt
streamlit
google-genai
pandas
matplotlib